pyproject.toml
[project]
name = "mqtt-teste"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "flask-socketio>=5.6.1",
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
]
[tool.pyright]
"typeCheckingMode" = "off"
[tool.ruff]
select = []
